skills/database-query-skill/pyproject.toml
[project]
name = "database-query-skill"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["psycopg2-binary", "pymysql"]
